requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
python-jose
passlib[argon2]
pydantic
